fix: add create-example-nft as example
