+++ scaffold-sui/.env.local
@@ -1,6 +1,6 @@
 NEXT_PUBLIC_NFT_STORAGE_KEY=
 NEXT_PUBLIC_WALLET_PRIVATE_KEY=
-NEXT_PUBLIC_DAPP_ADDRESS=0x65f4a0954aa6e68d2381ff98b7676df2fe57beee3ca37a4a8a57fa621c1db872
+NEXT_PUBLIC_DAPP_ADDRESS=0x0000000000000000000000000000000000000002
 NEXT_PUBLIC_DAPP_NAME=DID
 NEXT_PUBLIC_MARKET_FEE_NUMERATOR=10
 NEXT_PUBLIC_MARKET_INITIAL_FUND=10000
