+++ scaffold-sui/src/components/SuiConnect.tsx
@@ -7,7 +7,6 @@ import {
 
 export function SuiConnect() {
   const { account, connected, allAvailableWallets } = useWallet();
-  console.log(allAvailableWallets)
   return (
       <ConnectButton>Connect Wallet</ConnectButton>
   );
